Move LinkSection to components
