+++ packages/frontend/src/pages/scaling-projects/props/getLinkSection.ts
@@ -1,6 +1,6 @@
 import { Layer2 } from '@l2beat/config'
 
-import { LinkSectionProps } from '../view/links/LinkSection'
+import { LinkSectionProps } from '../../../components/project/links/LinkSection'
 
 export function getLinkSection(project: Layer2): LinkSectionProps {
   return {
